+++ plugin/brdrq/brdrq_module_importer.py
@@ -12,7 +12,7 @@ from PyQt5.QtWidgets import QMessageBox
 # https://github.com/qgis/QGIS/issues/45646
 
 
-brdr_version = "0.15.5"
+brdr_version = "0.15.6"
 
 
 def find_python():
@@ -44,7 +44,13 @@ def pipinstall_by_subprocess(python_exe,package):
         try:
             pipinstall_in_libs(python_exe,package)
         except:
-            QMessageBox.error(None, "Error", f"This plugin needs external dependency '{package}', automatically installed for Windows. For Linux/Mac, the correct version ({package}) has to be installed manually")
+            QMessageBox.critical(
+                None,
+                "Dependency Error",
+                f"This plugin needs external dependency '{package}'. "
+                f"Automatic installation is only supported on Windows. "
+                f"For Linux/Mac, install {package} manually.",
+            )
 
     else:
         try: